fix(database): transaction label exceeds sqlite uint64 (#2357)

+++ database/plugin/metadata/sqlite/transaction.go
@@ -566,7 +566,7 @@ func (d *MetadataStoreSqlite) GetTransactionsByMetadataLabel(
 
 	subQuery := db.Model(&models.TransactionMetadataLabel{}).
 		Select("transaction_id").
-		Where("label = ?", label)
+		Where("label = ?", types.Uint64(label))
 
 	query := db.
 		Where("id IN (?)", subQuery).
@@ -608,7 +608,7 @@ func (d *MetadataStoreSqlite) CountTransactionsByMetadataLabel(
 
 	var count int64
 	if result := db.Model(&models.TransactionMetadataLabel{}).
-		Where("label = ?", label).
+		Where("label = ?", types.Uint64(label)).
 		Count(&count); result.Error != nil {
 		return 0, fmt.Errorf(
 			"count txs by metadata label %d: %w",
@@ -967,7 +967,7 @@ func (d *MetadataStoreSqlite) SetTransaction(
 		for _, tmpLabel := range metadataLabels {
 			labelRecords = append(labelRecords, models.TransactionMetadataLabel{
 				TransactionID: tmpTx.ID,
-				Label:         tmpLabel.Label,
+				Label:         types.Uint64(tmpLabel.Label),
 				Slot:          point.Slot,
 				CborValue:     tmpLabel.CborValue,
 				JsonValue:     tmpLabel.JsonValue,
@@ -2494,7 +2494,7 @@ func (d *MetadataStoreSqlite) SetTransactionBatched(
 		for _, tmpLabel := range metadataLabels {
 			labelRecords = append(labelRecords, models.TransactionMetadataLabel{
 				TransactionID: tmpTx.ID,
-				Label:         tmpLabel.Label,
+				Label:         types.Uint64(tmpLabel.Label),
 				Slot:          point.Slot,
 				CborValue:     tmpLabel.CborValue,
 				JsonValue:     tmpLabel.JsonValue,
